backend/requirements.txt
google-cloud-pubsub
google-auth
google-api-core